requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
Cybernator==0.6.2
discord==1.0.1
discord.py==1.5.0
embed==0.1.0
idna==2.10
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
nekos.py==1.0.3
numpy==1.19.2
Pillow==7.2.0
psutil==5.7.2
pylint==2.6.0
pymongo==3.11.0
pynput==1.7.1
pyzbar==0.1.8
requests==2.24.0
six==1.15.0
speedtest-cli==2.1.2
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
yarl==1.5.1
